refactor: Hide Tab Bar Extension
