+++ frontend-ios/YakAl/YakAl/Sources/Views/Mypage/TimeSettingView.swift
@@ -21,7 +21,8 @@ struct TimeSettingView: View {
     @State private var isShowingAlert = false
     @State private var isShowingDatePicker: Bool = false
     @State private var isSettingStartTime: Bool = true  // Start with startTime by default
-
+    @Environment(\.viewController) private var viewControllerHolder: UIViewController?
+    
     var isValidTime: Bool {
         return startTime < endTime
     }
@@ -37,7 +38,22 @@ struct TimeSettingView: View {
         formatter.timeStyle = .short
         return formatter.string(from: endTime)
     }
+    // bar가리기
+    func hideTabBar() {
+        if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
+           let tabBarController = windowScene.windows.first?.rootViewController as? UITabBarController {
+            tabBarController.tabBar.isHidden = true
+        }
+    }
 
+    // bar보이게
+    func showTabBar() {
+        if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
+           let tabBarController = windowScene.windows.first?.rootViewController as? UITabBarController {
+            tabBarController.tabBar.isHidden = false
+        }
+    }
+    
     var body: some View {
         ZStack {
             VStack(spacing: 16) {
